(fix) Break Integration Tests into small sets of use cases
