+++ integration-tests/circuit-breaker/src/main/openapi/simple-openapi.json
