--- integration-tests/example-project/src/main/resources/application.properties
@@ -1,40 +1,0 @@
-# Authentication properties
-org.acme.openapi.security.auth.basic_auth/username=john
-org.acme.openapi.security.auth.basic_auth/password=test
-
-# since the file name has a space, we use the URI representation of this space here to not break the properties file
-# see the RFC3986 for more info https://datatracker.ietf.org/doc/html/rfc3986
-quarkus.openapi-generator.codegen.spec.open_weather_yaml.base-package=org.acme.openapi.weather
-quarkus.openapi-generator.codegen.spec.open_weather_yaml.custom-register-providers=org.acme.openapi.weather.api.provider.CountProvider
-quarkus.openapi-generator.codegen.spec.open_weather_no_security_yaml.base-package=org.acme.openapi.weathernosec
-# Authentication properties
-quarkus.openapi-generator.open_weather_yaml.auth.app_id.api-key=12345
-quarkus.openapi-generator.open_weather_no_security_yaml.auth.app_id.api-key=12345
-
-# KOGITO-6458 - generate auth bindings even if security definition is missing
-# Note: The property value is the name of an existing securityScheme in the spec file
-quarkus.openapi-generator.codegen.default.security.scheme=app_id
-
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/enabled=true
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/failOn = java.lang.IllegalArgumentException,java.lang.NullPointerException
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/skipOn = java.lang.NumberFormatException, java.lang.IndexOutOfBoundsException
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/delay = 33
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/delayUnit = MILLIS
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/requestVolumeThreshold = 42
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/failureRatio = 3.14
-org.acme.openapi.simple.api.DefaultApi/byeGet/CircuitBreaker/successThreshold = 22
-
-
-quarkus.openapi-generator.codegen.spec.multipart_requests_yml.base-package=org.acme.openapi.multipart
-quarkus.openapi-generator.codegen.spec.multipart_requests_yml.additional-model-type-annotations=@io.quarkus.runtime.annotations.RegisterForReflection
-# By default the openapi-generator doesn't generate models for multipart requests
-quarkus.openapi-generator.codegen.spec.multipart_requests_yml.skip-form-model=false
-
-quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.type-mappings.UUID=String
-quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.type-mappings.File=InputStream
-quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.import-mappings.File=java.io.InputStream
-quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.base-package=org.acme.openapi.typemapping
-
-quarkus.openapi-generator.codegen.spec.open_weather_customer_register_provider_yaml.base-package=org.acme.openapi.weathercustomregisterprovider
-quarkus.openapi-generator.codegen.spec.open_weather_customer_register_provider_yaml.custom-register-providers=org.acme.openapi.weather.api.provider.CountProvider
-
