+++ integration-tests/generation-input/README.md
@@ -1,0 +1,5 @@
+# OpenAPI Generator - OpenAPISpecInputProvider Example
+
+This module exemplifies a library that provides the OpenAPI specification file as an input stream reference.
+
+See [generation-tests](../generation-tests) as a how to use example.
