+++ integration-tests/generation-tests/README.md
@@ -1,0 +1,5 @@
+# OpenAPI Generator - OpenAPISpecInputProvider Example Usage
+
+This module exemplifies how to use an input stream instead of a file in the `openapi` directory to generate the Rest Clients.
+
+See the [generation-input](../generation-input) module as a library reference.
