+++ integration-tests/multipart-request/src/main/openapi/multipart-requests.yml
