+++ integration-tests/multipart-request/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+quarkus.openapi-generator.codegen.spec.multipart_requests_yml.base-package=org.acme.openapi.multipart
+quarkus.openapi-generator.codegen.spec.multipart_requests_yml.additional-model-type-annotations=@io.quarkus.runtime.annotations.RegisterForReflection
+# By default the openapi-generator doesn't generate models for multipart requests
+quarkus.openapi-generator.codegen.spec.multipart_requests_yml.skip-form-model=false
