+++ integration-tests/register-provider/src/main/java/org/acme/openapi/weather/api/provider/CountProvider.java
