+++ integration-tests/register-provider/src/main/openapi/open weather customer register provider.yaml
