+++ integration-tests/register-provider/src/main/openapi/open weather.yaml
