+++ integration-tests/register-provider/src/main/resources/application.properties
@@ -1,0 +1,7 @@
+# Custom Register provider use case with security
+quarkus.openapi-generator.codegen.spec.open_weather_yaml.base-package=org.acme.openapi.weather
+quarkus.openapi-generator.codegen.spec.open_weather_yaml.custom-register-providers=org.acme.openapi.weather.api.provider.CountProvider
+
+# Custom Register provider use case without security
+quarkus.openapi-generator.codegen.spec.open_weather_customer_register_provider_yaml.base-package=org.acme.openapi.weathercustomregisterprovider
+quarkus.openapi-generator.codegen.spec.open_weather_customer_register_provider_yaml.custom-register-providers=org.acme.openapi.weather.api.provider.CountProvider
