+++ integration-tests/security/src/main/resources/application.properties
@@ -1,0 +1,17 @@
+# Authentication properties
+org.acme.openapi.security.auth.basic_auth/username=john
+org.acme.openapi.security.auth.basic_auth/password=test
+
+# since the file name has a space, we use the URI representation of this space here to not break the properties file
+# see the RFC3986 for more info https://datatracker.ietf.org/doc/html/rfc3986
+quarkus.openapi-generator.codegen.spec.open_weather_yaml.base-package=org.acme.openapi.weather
+quarkus.openapi-generator.open_weather_yaml.auth.app_id.api-key=12345
+
+# Authentication properties
+quarkus.openapi-generator.codegen.spec.open_weather_no_security_yaml.base-package=org.acme.openapi.weathernosec
+quarkus.openapi-generator.open_weather_no_security_yaml.auth.app_id.api-key=12345
+
+
+# KOGITO-6458 (https://issues.redhat.com/browse/KOGITO-6458) - generate auth bindings even if security definition is missing
+# Note: The property value is the name of an existing securityScheme in the spec file
+quarkus.openapi-generator.codegen.default.security.scheme=app_id
