+++ integration-tests/simple/src/main/openapi/quarkus-simple-openapi.yaml
