+++ integration-tests/simple/src/test/java/io/quarkiverse/openapi/generator/it/QuarkusSimpleOpenApiTest.java
