+++ integration-tests/skip-validation/src/main/openapi/awx.json
