+++ integration-tests/skip-validation/src/test/java/io/quarkiverse/openapi/generator/it/WiremockAWX.java
