--- integration-tests/test-project/src/main/openapi/openapi3.0-security.yaml
@@ -1,48 +1,0 @@
----
-openapi: 3.0.3
-info:
-  title: Generated API
-  version: "1.0"
-paths:
-  /:
-    post:
-      operationId: doOperation
-      security:
-        - client_id: [ ]
-        - oauth: [ read, write ]
-        - bearerAuth: [ ]
-      requestBody:
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/MultiplicationOperation'
-      responses:
-        "200":
-          description: OK
-components:
-  schemas:
-    MultiplicationOperation:
-      type: object
-  securitySchemes:
-    client_id:
-      type: apiKey
-      in: header
-      name: X-Client-Id
-      x-key-type: clientId
-    bearerAuth:
-      type: http
-      scheme: bearer
-    oauth:
-      type: oauth2
-      flows:
-        authorizationCode:
-          authorizationUrl: https://example.com/oauth/authorize
-          tokenUrl: https://example.com/oauth/token
-          scopes:
-            read: Grants read access
-            write: Grants write access
-            admin: Grants read and write access to administrative information
-        clientCredentials:
-          tokenUrl: http://localhost:8382/oauth/token
-          scopes:
-            read: read
