+++ integration-tests/type-mapping/src/main/openapi/type-mappings-testing.yml
