+++ integration-tests/type-mapping/src/main/resources/application.properties
@@ -1,0 +1,4 @@
+quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.type-mappings.UUID=String
+quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.type-mappings.File=InputStream
+quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.import-mappings.File=java.io.InputStream
+quarkus.openapi-generator.codegen.spec.type_mappings_testing_yml.base-package=org.acme.openapi.typemapping
